find_unsafe2: refactor: move Visitor trait into new mir_visitor module
